Changes to environment handling and ast walks
The concept of an environment in introduced which holds references
to all of our symbols available to the executing code.

Reserved function def will add an element to the environment by its
textual representation with the following node as its value.

Reserved function ' will generate a literal list from a normal list.